refactor(react_refresh): split lib.rs into lib, scanner, helpers, and tests modules
Same shape as the earlier refactors. The 829-line crates/react_refresh/src/lib.rs
is split into four modules. All existing unit tests still pass.
